fix(fe): move standings table to component folder
